+++ dcb/src/main/java/org/olf/dcb/ingest/IngestService.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.Function;
 
+import org.olf.dcb.core.clustering.ImprovedRecordClusteringService;
 import org.olf.dcb.core.clustering.RecordClusteringService;
 import org.olf.dcb.core.model.BibRecord;
 import org.olf.dcb.core.svc.BibRecordService;
@@ -32,6 +33,7 @@ import reactor.core.Disposable;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.core.publisher.MonoSink;
+import services.k_int.features.Features;
 import services.k_int.federation.reactor.ReactorFederatedLockService;
 import services.k_int.micronaut.PublisherTransformationService;
 import services.k_int.micronaut.concurrency.ConcurrencyGroupService;
@@ -43,6 +45,10 @@ public class IngestService implements Runnable, ApplicationEventListener<Applica
 
 	public static final String TRANSFORMATIONS_RECORDS = "ingest-records";
 	public static final String TRANSFORMATIONS_BIBS = "ingest-bibs";
+	
+	public static int getProcessVersion() {
+		return Features.featureIsEnabled( ImprovedRecordClusteringService.FEATURE_IMPROVED_CLUSTERING ) ? 4 : 3;
+	}
 		
 	@Value("${dcb.shutdown.maxwait:0}")
 	protected long maxWaitTime;