+++ dcb/src/main/java/org/olf/dcb/processing/matching/MatchKey.java
@@ -1,8 +1,0 @@
-package org.olf.dcb.processing.matching;
-
-import java.util.UUID;
-
-public interface MatchKey {
-	public UUID getRepeatableUUID();
-	public String getText();
-}